gitserver: repoExists checks for HEAD file
Summary:
Previously we just checked if it was a valid directory, now we check if
it is a valid GIT_DIR.

Previously paths like `github.com` would of been regarded as `RepoExists`, but
now they will not. We likely need to follow up with error codes related to
invalid paths.

Depends on D76

Test Plan: CI. Will test all my gitserver changes on staging

Reviewers: richard

Reviewed By: richard

Differential Revision: https://srcgraph.phacility.com/D77