Exclude hash fragment-dependent state in Blob's initial rendering.
This needs to be done to support server-side rendering without creating
a conflict with frontend, since backend can't know the hash fragment
value in advance.

Defer actions based on hash fragment until after initial rendering and
attachment to DOM.

Use a timeout to defer initial scrollTo until startLine is populated by
BlobRouter. The alternative of propagating it from BlobRouter to
BlobContainer to Blog was not worth the time and code amount,
complexity. It can be revisited as needed later.