vcs: Handle vcs.ErrRepoExist
Summary:
There are races / pre-existing data which cause calls to Clone or Init
to fail with `vcs.ErrRepoExist`. In those cases we should treat the Clone/Init
function as idempotent, and ignore the error. The reason this is correct
behaviour is because the callers just want the repo to be there.

I could not reproduce the problem, but it is still happening according to
Sentry. So instead I followed the callpaths backwards from `gitserver.create`.

Fixes https://app.asana.com/0/37478073567611/102851029088772

Test Plan: CI

Reviewers: slavery, richard, beyang

Reviewed By: richard

Differential Revision: https://srcgraph.phacility.com/D91