app: always show repo cloning interstitial
Fixes the issue where if the repo clone succeeded in the timeout
period (1s), then the repo page would be initially blank.

Simplifies the handling of no-VCS-data errors to prevent this type of
issue in the future.

Now, only the repo homepage will trigger a RefreshVCS clone for
mirrors. Previously, any repo page would do so. That was not necessary
and led to increased complexity (since this RefreshVCS-if-needed call
could occur anywhere, instead of being localized to one place).