simplify def routes: enforce single path component (no "/") in def unit and def path URL path components
The underlying srclib data is not affected. This is a lossy frontend
translation, so that any "-" in a def unit or def path is converted to
a "/" when the def is fetched from the srclib store. Likewise, to
generate a URL for a def, any "/" is replaced with a "-".

This change is transitional; after all known srclib toolchains have
been modified to avoid emitting slashes in def units and def paths,
and when srclib enforces this, then we will remove the translation as
it will no longer be necessary.