simplify repo routes: use /my/repo/-/subroute not /my/repo/.subroute
Any path components up to the first "/-/" in the URL are treated
as part of the repo or rev. Anything after that is the subroute.

This makes it possible to use react-router to perform URL routing
on the frontend. It also greatly simplifies our regexps for routing.

Also, the BuildVarsFunc and PostMatchFunc used to perform repo rev URL
routing were removed, in favor of the simpler approach of adding the
"@" in the route helpers. This will enable us to use standard
gorilla/mux (not our fork) and makes routing implementation simpler.