Move in syntaxhighlight package into ./pkg/syntaxhighlight.
By having a copy of that package internally in /pkg/, we can more
easily break APIs and evolve it to our needs without being constrained
by API stability and backwards compatibility.

When the API becomes more stable and there's a demand, we can then have
a more stable public copy of it in an external repository.

This is a similar change to the go-vcs package move.