Factor out logic to select lexer for a given file into sourcegraph.
Previously, that logic was inside the general syntaxhighlight library.
That doesn't work well, since different users of syntaxhighlight may
have different strategies for highlighting files. What would happen
otherwise is that the general package would need to be changed to
specific needs of sourcegraph.

Instead, change signature of syntaxhighlight.Annotate to accept a
lexer. The lexer can be chosen and provided by the caller. That is a
much more appropriate and flexible API.