app: use node.js process to render React components, not cgo go-duktape
Now that we are rendering the entire page (in the upcoming
`pure-react` branch), general render perf is more important. The Go
fastpath for blob (which is only really possible with duktape) no
longer provides an advantage overall, since it is only used in some
cases.

This change makes the Go and JS behavior even more consistent, since
there is no longer any Go impl of the Blob component.