localstore: group db stores into two schemas
This commit modifies the gorp mapping in localstore to group db stores
into two separate schemas: AppSchema and GraphSchema. The schemas can
be configured to use the same or different pgsql database and/or host
based on environment variables. Currently all stores are grouped under
AppSchema, but the global_refs store will be grouped under GraphSchema.