httpapi: prevent against CSRF in logout
Tested using a simple HTML form like:

```
<html>
<body>
<form action="http://localhost:3080/.api/logout" method="post">
	<input type="submit" value="Submit">
</form>
</body>
</html>
```

Prior to this change, it would log you out. After this change, it shows an error:

```
cannot log out (no logged in user in context)
```

Additionally, normal logout was verified as working after this change.

We do not need additional CSRF prevention measures within httpapi, because we do not
allow any cookie-based authentication (see `httpapi/httpapi.go:29`). This specific CSRF
vulnerability was due to the fact that we would arbitrarily clear the user's session
regardless of whether or not they were authenticated (or not, in the case of a remote
site like the above HTML form).

Note that the lack of `handlerutil.UserMiddleware` additionally meant that:

- `serveSignup` could not detect logged in users, and thus would not prevent signed
  in users from creating new accounts (even though it explicitly tries to check for
  this).
- `serveForgotPassword` could not detect logged in users, and thus would allow
  requesting a forgot-password email while logged in.
- `servePasswordReset` could not detect logged in users, and thus would allow form
  submission of password change while logged in (this is the most scary, but I _think_
  because the token would have to be retrieved from the user's email account this could
  not be abused, I did not confirm this though).