server/local: gracefully handle errors in accessing global refs store
This is a temporary change to ensure that the app can handle errors in
accessing to the global refs store during the rollout. If a Sourcegraph
deployment does not have a secondary pgsql db available, this will
ensure that functionality on the site will fallback to the previous
behavior of displaying only local refs.

This commit should be reverted once the rollout is complete.