server/local: restrict external repo ref search to last built commit of the repo
This modifies the Defs.ListRefs endpoint to take a CommitID parameter,
which will restrict the ref search to that commit. If the parameter is
not specified, the queried def's commitID will be used. If that is not
specified either, the call will fail.

This ensures that the query to graphstore will always be restricted to
a single source tree, as otherwise it will fetch graph data for all
commits of the repo which is extremely slow at the moment.