httpapi: scope the local repo ref search to the def's commit ID
When falling back to graphstore for fetching ref locations in the def's
local repo, a commit ID must be specified to scope the ref search to
a single source tree. This change fixes the fallback behavior, so that
repos which haven't been built after xrefstore rollout will still
be able to serve local refs.