app/internal/ui: eliminate creation of unnecessary appdash parent span
This allows better tracing of the JS prerender time by including the
time underneath the app handler correctly, and including its sub-calls
in its span. Prior to this, there was an empty parent that caused these
calls to not be nested properly in the appdash data model (because an
empty parent does not allow appdash to connect the parent and children).