httpapi: add cache headers (with default max age) to some endpoints
This is generally a value like 60s in production. The occasionally stale
data (most likely to occur after a rebuild with new srclib data, for
annotations, or for a commit) is a tradeoff for greatly increased client
performance.

For comparison, GitHub uses a standard `Cache-Control: max-age=60` for
virtually all of its HTTP API endpoints.