pkg/vcs/gitcmd: cache commit log results
Uses an LRU cache as many other methods in this file do. Only caches
when the data is essentially mutable (except for oids being
deleted/GC'd).

The main goal of this is to speed up
Repos.GetSrclibVersionDataVersionForPath calls, which need to list
commits 1-2 times. It is not possible to cache at a higher level because
the existence of srclib data at a given commit can change at any time.

Caching it here makes GetSrclibVersionDataVersionForPath much faster
since they only need to incur GetCommit/ResolveRevision calls, which are
~1/8 as long as listing commits.