worker/plan: Collapse unsupported CI/lang warnings
For repos with many languages, the Build page is very crowded with yellow cells
saying we don't support the language. Instead, we can say it once, and list all
the unsupported languages in that message.

Test Plan: `src check` on kubernetes and srclib-go repos. Note that all the
unsupported languages in build and code intelligence are collapsed into one
warning.