Created a new handler to transform httpapi errors into a errorMessage struct, which then is parsed in the front-end and displayed to the user. Now, instead of simply telling the user "there has been a problem" when they do things like: type an incorrect password, try to register a new account with the login/email that has already been used, etc., the user is informed what they did wrong.
This change is made so that users are informed what went wrong. Before this change, sourcegraph
just displays "Sorry, there's been a problem.", which 1) does not tell user what went wrong,
and 2) makes it look like some kind of problem that happened in sourcegraph/our end. With this
change, it is very clear what the user needs to change (if indeed it is a user error).