app: ignore multiple-refs-incident-on-a-token, just go to the first def arbitrarily when clicking
Multiple refs incident on a single token occurs in Go when you have embedded fields:

type A struct {}

type B struct { A }

On the second "A" token, there is both a ref to the type "A" and the
field "(B).A".

We used to handle this by showing a disambiguation menu. However, that
broke easily because it was a rarely used feature. We removed the
feature but had a TODO thrown exception that made it so you couldn't
even click on the fields at all. This change makes clicking go to one
of the refs arbitrarily, which is better than doing nothing.

This workaround is fine for Go, but not fine for languages that have
more multiple-refs, such as Java and Go. When they are a focus, we
will bring back the disambiguation menu.