reuse GetSrclibDataVersionForPath resolution from fetching file/def when fetching annotations
Eliminates the duplicate "automagic" srclib-version-resolution in the
GET /.api/annotations endpoint, and reuses the resolution in the app JS
code (from the file or def fetches).

The resolution is an expensive operation to perform, so it's best to
reuse the result. It's also good to avoid magic.

The downside of this change is that the annotations fetch must wait for
the blob fetch to return in order to be fired, but we can optimize that
problem away shortly.