eliminate default branch lookup and more comprehensively highlight active def on refs main page
This commit allows for "fuzzy active def matching": if the active def
matches a link, or if the active def without a revision matches a
link, then the link is highlighted (e.g., on the refs page). This
allows us to highlight refs that appear in repos other than the def's
repo, on the def's ref's page.