app: increase render cache hit rate by making cache key only contain important state
Most of the data that was going into the cache key was not actually
part of the application state; the appdash spah ID, for example. That
meant that the cache was effectively useless because every request
would have a different span ID. This change eliminates those from the
cache key determination to make the cache effective again.