server/accesscontrol: Report more accurate errors for GitHub permissions.
Previous error message was indistinguishable from when you're not logged in
to Sourcegraph. This error message makes it clear that permission is denied
by GitHub API, so logging in to Sourcegraph will not help (you may already
be logged in to Sourcegraph).

Add note to explain why we must preserve unauthenticated error when we're
not sure if it's unauthenticated or unauthorized.