server/accesscontrol: Don't skip "github.com/" repo check for other capitalizations.
This change makes it very obvious that the if strings.HasPrefix(repo, "github.com/")
blocks will run in all situations where repo begins with "github.com/", even if
the capitalization of github is not lowercase.

Such situations should not occur since we have an invariant elsewhere in the
codebase preventing it from happening, but it's easier to add this enforcement
locally here rather than to prove that it won't happen externally.