app/web_modules/sourcegraph/app: (appdash) record route name for initial page load
This change causes the name `load page <route>` to be recorded instead of just
`page load` for the initial page load; for views the name remains as `load view <route>`
such that the naming is consistent.

Refactor relevant code into WithAppdashRouteStateRecording such that we have access
to the routes object in order to build the route name properly.