httpapi: attach appdash spans as siblings of the POST request span
This is partially what `6c00d2` removed, except it attached them as
children of the trace root (rather than as siblings to the POST request)
and thus could lose valuable context. By attaching it is a sibling span,
we keep valuable context and also do not hide them under the POST request
span.