app: escape raw JSON that is included in HTML templates
Previously, if the rawJSON template func's input (which, for preloaded
store data, comes from jsserver JSON.stringify) contains a string such
as "</script>", and the output is included in an HTML page, some
browsers (incl. Firefox) would interpret that as ending the outer
script tag. Any subsequent characters will be interpreted as a
continuation of the HTML document.

This fix properly escapes the JSON so that it is safe to include in an
HTML page.

Test plan:

1. Visit http://localhost:3080/%3Cscript%3Ealert%28123%29%3C/script%3E in Firefox. Ensure that no alert is printed. Note: you mut have the JS prerenderer enabled to trigger the alert, so ensure you aren't setting the env var SG_DISABLE_JSSERVER=t when you run this test plan.