server/local: Resolve renamed remote repos correctly.
If the remote repo has a different canonical path, check if it already
exists locally before saying it's a remote repo. Otherwise, the
frontend will think it needs to be cloned (which may or may not be the
case) and try to do that, which will fail with 400 bad request if repo
already exists.

This way, Resolve works as its name suggests and fully resolves repos
in this situation.