app: fix and improve positioning of def box
Reduces complexity of the positioning of the def box in the file
margin. Removes the need for extra padding computations, complex pixel
offset callbacks, etc. Enables the def box to be server-side rendered.

The one complexity is that certain full-page zoom levels (90%, 110%)
that do not cleanly divide the 16px code line height produce slightly
incorrect offsets. Therefore we need to compute the actual height of a
16px-height box (which is, e.g., 15.556px at 90% full-page zoom, due
to Chrome's/etc.'s internal positioning algorithms).

The def box can now be prerendered and prepositioned on the server,
which enables better caching and reduces page load jitter. Previously
we were doing all the work to render the def box on the server but not
actually including it in the response, so this will not increase load
on the server or TTFB.