app: instant navigation for jump-to-defs in the same file
Previously, even for defs in the same file, jump-to-def would have to
load the def from the server before jumping to it in the same file. It
is very cheap to locally determine that a ref's target is a def in the
same file, so that network roundtrip was unnecessary.

This fixes the issue where clicking on refs to a def in the same file
would cause a full page transition and gray shimmer.

Clicking on something defined in the same file now NEVER causes a page
reload, and it jumps to the def instantly (with no network fetch
required). This is hard to capture in our (or any) perf measurements
but it vastly increases the perceived speediness of the app. The add'l
cost to offer this feature is about 1% increase in annotations byte
size over the network and client-side ~1-5msec of indexing the
annotations for a given file. It does not introduce any delay when
jumping to defs defined in another file.