app: scroll to things that are on the screen but near the edges, to avoid accidentally missing them
If you click on a ref to a def that is right near the bottom of your
screen, previously it wouldn't scroll you down. So, you might not
notice that anything got highlighted. This change makes it so anything
within 150px of the top or bottom of the screen is considered
off-screen for the purposes of jump to def.

If you click a def near the bottom of the screen, it will actually
scroll a whole page so that the thing you clicked is at the top of the
screen. This the naive implementation and a smarter behavior is a bit
more involved.