app: remove left/right-arrow key actions for TreeSearch (they override default OS text insertion cursor actions)
Fixes the following issue.

Repro: go to https://sourcegraph.com/github.com/golang/go and type in
Buffer. Click the mouse to put the text insertion cursor to the left
of the B in the search input. Now hit the right arrow key to move your
text insertion cursor.

Actual: it goes to the first result
Expected: it just moves the text insertion cursor