web_modules: Handle zero annotations without panicking.
Fixes https://app.asana.com/0/87040567695724/115091483542820.

Don't add unnecessary safety guards now, because it will make it hard,
if not impossible, to remove them in the future. To increase our chances
of not accidentally having type errors, we can make better use of structs
with types in JavaScript so type checked will be able to catch those issues.