localstore: optimize SQL query in GlobalRefs.Get
The older version of the query would first do a scan and sort and
then limit the result. This resulted in queries for popular defs to
scan a large number of rows and slow down the request.

The update query will first apply a limit and then do the sorting
and aggregation. This means that the result may not contain the
global top repos, but will be a sample of the repos referring a def.