Update to Go 1.6.2.
This change updates from Go 1.6.1 to 1.6.2, a minor point release.

This release includes fixes to the compiler, runtime, tools,
documentation, and the mime/multipart, net/http, and sort packages. See
https://golang.org/doc/devel/release.html#go1.6.minor for details.

One highlight is a last-minute bug in net/http is fixed, which was
preventing the promised HTTP/2 to be on by default for HTTP clients.
Now, it will be on by default for HTTP clients. HTTP/2 is considered
stable, but this is a change worth keeping an eye out for.