graphstore: Cache __versions and idx for cloudstorage
After some investigation, we hit these sort of files a lot. `__versions` is
mutable, so we put a short lived cached for it since it can be hit several times
for a single request. `idx` files are the jumping point to reading the `def`
files. After some experimentation with this in production, it may make sense to
also cache the `dat` files.

Alternatively we could make the HTTP transport that gcloud uses speak to the
cache. However, gcloud only sets ETags and no other cache control headers. So
we would not avoid the latency of speaking to gcloud. This is why we took this
approach, to avoid the latency.

Test Plan: Ran a local server which spoke to remote. Added logging to see access
patterns. After re-loading a page, ensured we had no rerequests on `idx` or
`__versions` dir.

```
SRC_FED_IS_ROOT=true SRC_GRAPHSTORE_ROOT=gcloudstorage://graphstore-staging SRC_NO_WORKER=true GORACE= make serve-dev SRCFLAGS="--log-level dbug"
```