app: reset all data and reload views when auth changes (login/logout/signup)
When the user signs in, logs in, or logs out, all data needs to be
flushed and all views need to be redisplayed. (We can get more granular
with this in the future, but since we don't know which data is
user-specific and which isn't, we need to be coarse now.)

To do this, we need to actually reset the stores and trigger an
emitChange on each of them. We also need to tell the views to re-fetch
data when it is no longer available. Previously, we only needed to fetch
data when the route params (or derived values) changed; now we must also
fetch data when previously fetched data has been cleared.