app/web_modules/sourcegraph/def: make DefInfo pagination show only 4 prev/next page numbers
Prior to this change the page woudl show pagination links (numbers) for every
previous and next page. After this change it only shows them for up to 4, such
that there are at max 9 page links on the page with the center one being the
page you are on (which is disabled / unclickable).