app: resolve tree search to latest built commit
Before, the tree search would always link to results on the current rev.
If the current rev is a branch that is not built, its def link would
result in a 404. Now it will link to the latest built commit if it was
behind.

Another fix: xrefs was linking to defs on the rev of the current repo,
fixed this so it links using the rev of external def.