e2etest: Export prometheus metrics around test success
We will mainly use this to alert on e2etest not running. The state per test
counters are more useful to get some statistics on our test runs, and not for
alerting. However, it could be a way to do alerts in future.

Test Plan: Visit http://localhost:6060/metrics and eye-ball introduced metrics.
Ran

  $ PROMETHEUS_IO_ADDR=:6060 SELENIUM_SERVER_IP=$(docker-machine ip docker-vm) TARGET=https://sourcegraph.com e2etest -once=false