app/web_modules/sourcegraph/init: fix Splunk currentUser fetching
`ev["User"]` works because we already set it via `app/templates/scripts.html`
in order for it to end up in Splunk in the same mannor that we pass other build
info (like deployed build version, Appdash trace ID, etc.). `context.currentUser`
in contrast, no longer works or is correct.