pkg/vcs: compute file size in FileInfos returned by ReadDir, Stat, and Lstat
Fixes incorrect primary language determination (which was caused by the
"total number of bytes" in each programming language being calculated as
0 by pkg/inventory due to ReadDir always returning os.FileInfos with
Size() == 0).

As an optimization, ReadDir was previously changed to NOT compute the file
sizes (by omitting the `--long` flag from the `git ls-tree` call). This
is faster in cases when the file sizes are not needed, such as showing
the repo or tree pages in the app. However, that violates the
vfs.FileSystem ReadDir contract, and it was breaking the primary
language determination. If we add that optimization back, we should be
able to configure whether it is used, instead of always skipping file
size computation.