app: set Branch and Tag on newly created build in "Rebuild" action
The build's Branch tells drone's clone step to fetch the last 50
commits on that branch. But when you click Rebuild, it creates a new
build with the same commit ID but NO branch. Then drone's clone step
tries to fetch the last 50 commits from master, and if this build's
commit is not on there (which occurs for commits not on master, so
quite commonly), so it fails.