app: fix def info paginator not showing initially
Before, we only displayed the paginator button if the number of files
loaded from ref locations is >= to the page limit. This did not always
work since the bakend will load N locations per page from the DB, but it
will filter out some files based on repo permissions as well. Now we
check if there are more repos to load files from as well.