app/web_modules: Set canonical URL for repo pages.
Use the repo URL without revision specified as the canonical URL for
same pages with the default branch revision explicitly specified (i.e.,
@master). Other branches or commits continue to have their own url as
the canonical url.

Assume that default branch name is always "master". This may not always
be true, but it is for most git repos. Try this first, since it'll be
accurate for most repos, and figure out the actual default branch later
as a followup.