cli: Fix ?go-get=1 handling of mirror repos.
Previously, the handler tried to determine if a given repository is
a hosted one by checking if it exists. However, a repo may exist but
have Mirror flag set to true, in which case it's not a hosted repo.

This likely came to be after we've ran scripts to clone many repos.

Add test cases.

Fixes https://app.asana.com/0/87040567695724/134411614707569/f.