app: link sitemap index to static sitemaps hosted on Google Cloud Storage
This causes our sitemap index file to link to 9 static sitemaps that are hosted
on Google Cloud Storage; each consisting of 50,000 URL entries of def info routes
for a total of 450,000 def info pages that Google can index. More static sitemap
files or proper sitemap generation can be added at a later time. (this is a stopgap
solution).