app: add repo revision switcher to top global nav
Next to the filename or repo, there is a small downward-pointing
caret. It expands to a revision switcher. This revision switcher lets
you search all the branches/tags, shows you the branches (most
recently pushed first), tells you if the branch is behind the index,
and lets you switch to any of them. Switching to a branch preserves
the rest of the URL, so you'll see the same thing in the revision you
choose.

Sometimes what you're looking at doesn't exist in another revision. In
that case, it 404s. Better 404 pages will help with this experience,
but this is fairly easy to understand for files, and if you encounter
a 404 on a def (it's not built on the rev you switched to), it's easy
to switch back.

Test plan:

1. Use the rev switcher on blob, tree, repo, def-info, def, etc.,
   pages. Try revs in the URL like "master~3" as well as branches, empty,
   and full commit IDs.