global_refs: Force ref repo to be first
Due to the way we index and query the db, the order of results returned by
`GlobalRefs.Get` are completely dependent on the order the db wants to return
them. We could add an order by query on `innerSelectSQL`, but that would require
a complete scan and sort of all refs for a def which is expensive for defs with
a lot of refs (eg `fmt.Errorf`).

The most basic boosting we want is the repo for a ref be the first result shown
to a user. We have no ordering on `innerSelectSQL`, leading to us having to do
two queries. Luckily we do have an index on the `repo` column for Update.
As such we issue a query with `repo=op.Def.Repo` and `repo!=op.Def.Repo`
on the first page, and adjust pagination accordingly.

This approach means we will only ever return `opt.PerPageOrDefault` for the Def
Repo, but this should be fine since we have that set at `1000`. We will also get
to remove the slow kludge to query `Defs.ListRefs` if we don't have the Def
Repo, which seems to be the largest contributor to our p95 times. The other
trade-off is that we will return more than `opt.PerPageOrDefault` on the first
page. To not do this will require us to always calculate the count of
`op.Def.Repo` on other pages, which isn't worth the perf penalty.

Note that this is a simple serial implementation, but a parallel one should be
possible. Also note that this is only implemented for the new global refs
schema. We will be switching to the new schema very soon + the new schema seems
to suffer more from not including the ref's repo which lead to poor performance
previously.