app/web_modules: Add validation of first 2 arguments of autotest.
When possible (i.e., fs is available), read the contents of filename
argument and ensure it matches the expected argument. If they don't,
it could be because of invalid usage - they should match.

Hopefully this will help with usage of autotest, since we can't quite
simplify the two arguments into a single one because filesystem is not
available in all contexts where we might want to run these tests.

> This approach does not work when running the mocha tests in Chrome
> instead of via the mocha CLI. Chrome can not load files. While I
> was working on the JS codebase, running the tests in Chrome
> and using its nice JS debugger was very useful to me.

Closes #1955.