app: remove activeDefNoRev hack that's no longer needed
After 6cd74cc9f4 (the rev switcher), the rev used for annotating
same-file revs is consistent, so we no longer need to pass 2 variants of
the activeDef URL to ensure it is highlighted correctly.

This eliminates some expression evaluations and one property, which
should have a slight improvement in def hover perf. It also removes
now-needless complexity.