global_refs: Update from Graphstore rather than import
Our old approach had quite a few drawbacks
* It made srclib import slower (wasted cycles on workers)
* If a file was deleted, it would remain in global_refs
* Retrying an index involved retrying the whole build
* We had to massage raw srclib import data
* Reindexing the world required rebuilding the world

This new approach works similarly to the def store. When a new import srclib
import has finished, we background an update of the index. The data we index is
the same we return for local graphstore operations, and we can atomically update
all references for a repo in one go.

This approach is not perfect yet. We need two more additions (which would
benefit global_defs as well):

* Transactional queue of updates (so we don't lose indexing during a deploy)
* Incremental tuning (updating all refs in a large repo is a large transaction)

Fixes https://app.asana.com/0/87040567695724/133362528014721 and
https://app.asana.com/0/111205418767459/135279890598564

Test Plan: Unit tests as well as running builds on larger repos on staging.