app: enable scroll position updating
Previously, if you were (e.g.) halfway scrolled down a file on the
blob route, and then clicked on the homepage, you would also be
halfway scrolled down on the homepage. This is because browser scroll
behavior only takes over when there are full page transitions, not
react-router pushState/popState/replaceState transitions.

Now, the behavior is identical to that which you would expect,
including when you use the back/forward buttons.