app: fix handling of srclibDataVersionPath
Previously, this was not being set reliably. The only visible issue is
that it failed React propTypes validation. However, it could indicate
the presence of other bugs, such as not properly looking back to the
last-built commit.

The correct behavior is to look up the last srclib data version for the
path. The path is either determined statically from the URL (for blobs)
or after fetching the def and consulting def.File (for defs). This
change makes it so that both cases are handled properly.

Test plan:

1. Go to a def page with no rev, branch rev, and commit ID rev.
2. Same, but for a blob page.