e2e: Make test runner responsible for slack image upload
We had logic of writing an image in several places + the runTest method should
not care about reporting to a user, just collecting info. This change refactors
this portion of the code a bit to improve on this. It also prevents quarantined
tests from uploading images to the main #e2etest channel.

Fixes https://app.asana.com/0/87040567695724/139065576365330