app/web_modules/sourcegraph/tree: make TreeSearch properly requery SrclibDataVersion if there was no prior build
Prior to this change, in the event that a repository was never built (e.g. it was just
cloned) we would attempt to query the srclib data version _just once_ and if that 404'd
(because the repository hasn't been built yet) then we would not query ever again. This
would cause the `Sourcegraph is analyzing your code —- results will be available soon!`
message to display until you manually refresh the page. For small repositories, we didn't
notice this because the build typically finished before the user refreshed after seeing
the `Cloning this repository, please refresh in a minute` screen. However, for large
repositories whose builds take a long time it is much more noticable of an issue. Also,
now that we auto-refresh after clone it becomes an immediate issue for small repositories
too. This change fixes the issue by periodically polling for SrclibDataVersion if we
don't already have a valid one.