{api,app,services}: add RepoResolution.CanonicalPath with the canonical repo path
For now this is always the same as the Repo field. In the future the
Repo field will become the numeric ID, at which point having the
canonical path in a separate field will be necessary. This change is
being made now so that the future numeric ID switch will have less of
an impact.