builds: Lower bound on DequeueMsec (#45)
If there are no builds, no point asking straight away if there are new ones.
Instead of sleeping `[0,DequeueMsec]`, we now sleep
`[DequeueMsec*0.5, DequeueMsec*1.5]`. We also increase the default `DequeueMsec`
to 5s. For both production and development this is fine. In the case of
production we have many workers polling at independent times, so a larger sleep
interval won't affect speediness of picking up work. In dev a potential 7.5s
wait is acceptable.