github: Cache unauthed not found responses
This is by far our largest contributor to our rate limit usage, and we do no
caching on it at all (including at the transport layer since there is no
Cache-Control headers for 404s).

For example when I recently ran the splunk query

  sourcetype="*_prod_*" log="*TRACE github*" | rex field=log "path=(?<path>[^ ]+) code=(?<code>\d+)" | convert num(code) | stats count by code

for an hour it showed 10x more 404 responses than 200 (due to 200 being cached
correctly).

To confirm 404s contribute to the rate limit I ran the following command twice
and noticed the limit decreasing

  curl -si https://api.github.com/repos/sourcegraph/keegan-test

This change complicates the logic for caching. A follow-up commit will attempt
to simplify this core part of our reliability.

Note: We change what we serialize to the cache. But we do it in such a way it is
backwards compatible (I explicitly tested this), so we will not blow the cache.