services/ext/github: Move local error check up.
Perform githubutil.SplitRepoURI first thing in github.Repos.Get, before
even hitting the cache.

The cache is redis and may be a network call, etc., so the assumption
is that githubutil.SplitRepoURI is a much cheaper/quicker operation. If
the local parsing of githubutil.SplitRepoURI fails, there's no way it's
a valid repo that would be cached.

This is a minor optimization that can help in a situation where an
invalid repo URL is requested at a high rate.