builds: Prevent adding docker images prod does not support
In production we only support docker images which are v1 if we specify the
sha256 hash. If you build an image with docker v0.10 or higher, it breaks
srclib in production. This commit adds a test which speaks to the docker
registry to ensure the image exists and is an old enough version or uses a new
style tag.

This was tested by using the image in ff1c35f6a9796e7117623848234ea96a10c4d4f3
which introduced a bad image. Also we manually modified a sha to be non-existant
to test 404.