app: remove query param removal redirect after github oauth
This fixes "Logged in navigation all leads to code search"

There was an issue where after logging in via GitHub OAuth, React was
trying to attach event listeners to elements that didn't exist on the
page anymore, which caused a number of errors. The reason the elements
didn't exist is because they were <a> tags that belonged to the
AnonymousLandingPage component, which was rendered due to a reidrect
that happens after returning from a GitHub auth for the purpose of
removing the query params. This redirect naviagated to the home page,
but shortly after the user is redirected to the search page and there is
a conflict between the homepage being unmounted and the search page
rendering.