app: check that Container fields exist before trying to release/remove them
It appears that some React tooling (webpack dev server, hot reloading,
redbox) manually tears down components, apparently resulting in calls
to `componentWillUnmount` that did not have a corresponding
`componentDidMount` call. (I verified this by adding `console.count`
calls to both methods and observing that `componentWillUnmount` was
called but `componentDidMount` was not called when there was a redbox
error showing up.)

This results in another error showing up when
`this._containerStoreGroup.release()` is called on an undefined
`this._containerStoreGroup`. These errors confuse people and waste
time as people spend time looking into them, instead of the real error
that occurred.

This change will suppress those misleading errors and should not
affect production code.